+++ README.md
@@ -42,11 +42,11 @@
 
 ### [할 일 필터링]
 
-- [ ] **필터링 버튼 추가**
+- [x] **필터링 버튼 추가**
   - 할 일 목록 상단에 '전체', '진행 중', '완료' 버튼을 추가합니다.
-- [ ] **필터링 기능 구현**
+- [x] **필터링 기능 구현**
   - 사용자는 이 버튼들을 클릭하여 필터링할 수 있습니다.
-- [ ] **실시간 필터링 반영**
+- [x] **실시간 필터링 반영**
   - 필터링된 목록이 화면에 반영됩니다.
 
 ### [해야 할 일의 총 개수 확인]
